pmc/requirements.txt
kafka-python==2.2.10
tqdm==4.67.1
